+++ stacks/api.ts
@@ -49,10 +49,10 @@ export function API({ stack }: StackContext) {
         function: 'packages/functions/src/stripe/checkout.handler',
       },
       'GET /search/suggest': {
-        function: 'packages/functions/src/search/autosuggest.handler',
+        function: 'packages/functions/src/search/sports-suggest.handler',
       },
       'GET /money/search/suggest': {
-        function: 'packages/functions/src/search/autosuggest.moneyHandler',
+        function: 'packages/functions/src/search/money-suggest.handler',
       },
       'POST /stripe/manage': {
         authorizer: 'simple',
